Update storage-cleanup files with new architecture
